+++ packages/contracts/src/gateway/event/request/index.ts
@@ -1,0 +1,5 @@
+export * from './create-event.request.js';
+export * from './update-event.request.js';
+export * from './search-events.request.js';
+export * from './manage-requirement.request.js';
+export * from './tag.request.js';