+++ packages/contracts/src/gateway/event/request/tag.request.ts
@@ -1,0 +1,7 @@
+import type { CreateTagCommand, UpdateTagCommand } from '../../../event/event.command.js';
+
+/**
+ * Tag related requests
+ */
+export interface CreateTagRequest extends CreateTagCommand {}
+export interface UpdateTagRequest extends Partial<UpdateTagCommand> {}